pyproject.toml
[project]
name = "sparv-sbx-retreat2024-plugin"
version = "0.1.0"
description = "Uppercase converter (example plug-in for Sparv)"
readme = "README.md"
license.text = "MIT License"
# authors = [
#     { name = "", email = "" }
# ]
dependencies = [
    "sparv-pipeline~=5.0"
]
entry-points."sparv.plugin" = { sbx_retreat2024_plugin = "sbx_retreat2024_plugin" }